Move CodableFeedStore to its own production file
